Edit PostService

+++ src/main/java/br/com/projectblog/services/PostService.java
@@ -2,27 +2,30 @@ package br.com.projectblog.services;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
+import org.apache.tomcat.util.codec.binary.Base64;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
+import br.com.projectblog.domains.Role;
 import br.com.projectblog.dtos.PostDTO;
 import br.com.projectblog.dtos.UserDTO;
 import br.com.projectblog.dtos.filters.FilterPostDTO;
 import br.com.projectblog.dtos.requests.PostRequestDTO;
+import br.com.projectblog.exceptions.BusinessException;
 import br.com.projectblog.exceptions.ResourceNotFoundException;
 import br.com.projectblog.mappers.CommentMapper;
 import br.com.projectblog.mappers.PostMapper;
 import br.com.projectblog.mappers.UserMapper;
 import br.com.projectblog.models.Post;
 import br.com.projectblog.repositories.PostRepository;
+import br.com.projectblog.utils.UserUtils;
 import br.com.projectblog.validations.groups.OnCreate;
 import br.com.projectblog.validations.groups.OnUpdate;
 import jakarta.validation.Valid;
@@ -46,6 +49,9 @@ public class PostService {
 	
 	private final PostMapper mapper;
 	
+	public Optional<Post> findByIdPost(@NotNull UUID id) {
+		return this.repository.findById(id);
+	}
 	
 	public Optional<PostDTO> findById(@NotNull UUID id) {
 		Optional<Post> optPost = this.repository.findById(id);
@@ -66,6 +72,7 @@ public class PostService {
 				.userId(post.getUser().getUserId())
 				.title(post.getTitle())
 				.description(post.getDescription())
+				.image(post.getImage())
 				.commentsDTO(this.commentMapper.listCommentToListCommentDTO(post.getComments()))
 				.dateCreate(post.getDateCreate())
 				.dateUpdate(post.getDateUpdate()).build();
@@ -78,8 +85,13 @@ public class PostService {
 	
 	
 	public void deleteById(@NotNull UUID id) {
-		this.findById(id).orElseThrow(
+		PostDTO postDTO = this.findById(id).orElseThrow(
 				() -> new ResourceNotFoundException(String.format("Post pelo id %s não existe.", id)));
+		
+		UserDTO userDTO = this.userService.findById(postDTO.getUserId()).orElse(null);
+		
+		if(!UserUtils.getUsernameLogado().equals(userDTO.getUsername()) && !userDTO.getRole().equals(Role.ADMIN))
+			throw new BusinessException("Não é possível excluir post de outro usuário, a menos que seja o administrador.");
 		
 		this.repository.deleteById(id);
 		log.info("Post pelo id {} foi excluído com sucesso.", id);
@@ -95,10 +107,13 @@ public class PostService {
 			new ResourceNotFoundException(String.format("Usuário pelo identificador %s não encontrado.", postDTO.getUserId())
 		));
 		
+		if(!UserUtils.getUsernameLogado().equals(userDTO.getUsername()))
+			throw new BusinessException("Não é possível incluir post com outro usuário que não seja o seu.");
+		
+		postDTO.setImage(Base64.decodeBase64(dto.getImageEncodeBase64()));
 		postDTO.setDateUpdate(LocalDateTime.now());
 		postDTO.setDateCreate(LocalDateTime.now());
 		postDTO.setCommentsDTO(List.of());
-		postDTO.setImage(Base64.getDecoder().decode(dto.getImageEncodeBase64()));
 		
 		Post entity = this.mapper.postDTOToPost(postDTO);
 		entity.setUser(this.userMapper.userDTOtoUser(userDTO));
@@ -123,6 +138,9 @@ public class PostService {
 		UserDTO userDTO = this.userService.findById(UUID.fromString(dto.getUserId())).orElseThrow(() -> 
 			new ResourceNotFoundException(String.format("Usuário pelo identificador %s não encontrado.", dto.getUserId())
 		));
+		
+		if(!UserUtils.getUsernameLogado().equals(userDTO.getUsername()) && !userDTO.getRole().equals(Role.ADMIN))
+			throw new BusinessException("Não é possível atualizar post de outro usuário, a menos que seja o administrador.");
 		
 		postDTO.setTitle(dto.getTitle());
 		postDTO.setDescription(dto.getDescription());
